refactor: use vitest
