+++ tests/packages/integration/dependency-injection/mock/TestManager.js
@@ -1,0 +1,5 @@
+export class TestManager {
+    getName() {
+        return 'foobar';
+    }
+}
