+++ tests/packages/integration/dependency-injection/mock/test.di.yaml
@@ -1,0 +1,4 @@
+services:
+    TestManager:
+        from: "./TestManager"
+        import: TestManager
