+++ tests/packages/unit/dependency-injection/container/Container.test.js
@@ -1,5 +1,5 @@
-import {assert} from "chai";
 import {Container} from "@enhavo/dependency-injection/container/Container";
+import { describe, expect, test } from "vitest";
 
 class Callable
 {
@@ -22,14 +22,12 @@ class TestContainer extends Container
 }
 
 describe('dependency-injection/Container/Container', () => {
-    describe('test call function', () => {
-        it('should return value of called function', () => {
-            let container = new Container();
-            let callable = new Callable();
+    test('test call function should return value of called function', () => {
+        let container = new Container();
+        let callable = new Callable();
 
-            container._call('setName', callable, ['foobar']);
+        container._call('setName', callable, ['foobar']);
 
-            assert.equal('foobar', callable.getName())
-        });
+        expect.assert.equal('foobar', callable.getName())
     });
 });
