+++ tests/packages/unit/dependency-injection/mock/CompilerPass.js
