+++ tests/packages/unit/vitest.unit.js
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+    test: {
+        globals: true,
+        environment: "jsdom",
+        coverage: {
+            provider: 'v8' // or 'istanbul'
+        },
+    },
+})
